Task 2 finished

+++ Homework 3/old_src/seed/algorithms/ShortestPathAlgorithm.java
@@ -3,7 +3,7 @@ package seed.algorithms;
 public abstract class ShortestPathAlgorithm {
 
 	String name;
-	int timeComplexity;
+	long timeComplexity;
 	
 	abstract long findShortestPath(Graph g);
